Move integration tests. Write new integration tests
