+++ tests/v1/integration/test_integration_repseudonymize.py
@@ -1,16 +1,14 @@
-from collections.abc import Generator
-
 import polars as pl
+import pytest
 from polars.testing import assert_frame_equal
 
 from dapla_pseudo import Repseudonymize
-from tests.integration.utils import integration_test
-from tests.integration.utils import setup
+from tests.v1.integration.utils import integration_test
 
 
+@pytest.mark.usefixtures("setup")
 @integration_test()
 def test_repseudonymize_from_default_encryption_to_fpe(
-    setup: Generator[None, None, None],
     df_personer_fnr_ff31_encrypted: pl.DataFrame,
     df_personer_fnr_daead_encrypted: pl.DataFrame,
 ) -> None:
@@ -25,9 +23,9 @@ def test_repseudonymize_from_default_encryption_to_fpe(
     assert_frame_equal(result, df_personer_fnr_ff31_encrypted)
 
 
+@pytest.mark.usefixtures("setup")
 @integration_test()
 def test_repseudonymize_change_keys(
-    setup: Generator[None, None, None],
     df_personer_daead_encrypted_ssb_common_key_2: pl.DataFrame,
     df_personer_daead_encrypted_ssb_common_key_1: pl.DataFrame,
 ) -> None:
@@ -42,9 +40,9 @@ def test_repseudonymize_change_keys(
     assert_frame_equal(result, df_personer_daead_encrypted_ssb_common_key_2)
 
 
+@pytest.mark.usefixtures("setup")
 @integration_test()
 def test_repseudonymize_from_sid_to_daead(
-    setup: Generator[None, None, None],
     df_personer: pl.DataFrame,
     df_personer_daead_encrypted_ssb_common_key_1: pl.DataFrame,
     df_personer_daead_encrypted_ssb_common_key_2: pl.DataFrame,
